breaking: Migrate to Client from Bot
